🐛 fixed: Improved file structure
